chore(organizations): restructure controller exports and add unit tests for organization creation and deletion
