+++ modules/services/api/src/index.ts
@@ -1,3 +1,3 @@
 export { AppModule } from './app.module';
-export { OrganizationsController } from './org/controllers/organizations.controller';
+export { OrganizationsController } from './org/controllers/organizations';
 export { OrganizationsService } from './org/services/organizations.service';
